% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{AggregateBucket}
\alias{AggregateBucket}
\title{Fitness Objects 
Stores and accesses user data in the fitness store from apps on any platform.}
\usage{
AggregateBucket(activity = NULL, dataset = NULL, endTimeMillis = NULL,
  session = NULL, startTimeMillis = NULL, type = NULL)
}
\arguments{
\item{activity}{Available for Bucket}

\item{dataset}{There will be one dataset per AggregateBy in the request}

\item{endTimeMillis}{The end time for the aggregated data, in milliseconds since epoch, inclusive}

\item{session}{Available for Bucket}

\item{startTimeMillis}{The start time for the aggregated data, in milliseconds since epoch, inclusive}

\item{type}{The type of a bucket signifies how the data aggregation is performed in the bucket}
}
\value{
AggregateBucket object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:12:46
filename: /Users/mark/dev/R/autoGoogleAPI/googlefitnessv1.auto/R/fitness_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AggregateBucket Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

