% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{AggregateBy}
\alias{AggregateBy}
\title{AggregateBy Object}
\usage{
AggregateBy(dataSourceId = NULL, dataTypeName = NULL)
}
\arguments{
\item{dataSourceId}{A data source ID to aggregate}

\item{dataTypeName}{The data type to aggregate}
}
\value{
AggregateBy object
}
\description{
AggregateBy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The specification of which data to aggregate.
}
