% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{AggregateRequest}
\alias{AggregateRequest}
\title{AggregateRequest Object}
\usage{
AggregateRequest(aggregateBy = NULL, bucketByActivitySegment = NULL,
  bucketByActivityType = NULL, bucketBySession = NULL,
  bucketByTime = NULL, endTimeMillis = NULL,
  filteredDataQualityStandard = NULL, startTimeMillis = NULL)
}
\arguments{
\item{aggregateBy}{The specification of data to be aggregated}

\item{bucketByActivitySegment}{Specifies that data be aggregated each activity segment recored for a user}

\item{bucketByActivityType}{Specifies that data be aggregated by the type of activity being performed when the data was recorded}

\item{bucketBySession}{Specifies that data be aggregated by user sessions}

\item{bucketByTime}{Specifies that data be aggregated by a single time interval}

\item{endTimeMillis}{The end of a window of time}

\item{filteredDataQualityStandard}{A list of acceptable data quality standards}

\item{startTimeMillis}{The start of a window of time}
}
\value{
AggregateRequest object
}
\description{
AggregateRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Next id: 10
}
\seealso{
Other AggregateRequest functions: \code{\link{users.dataset.aggregate}}
}
