% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{AggregateResponse}
\alias{AggregateResponse}
\title{AggregateResponse Object}
\usage{
AggregateResponse(bucket = NULL)
}
\arguments{
\item{bucket}{A list of buckets containing the aggregated data}
}
\value{
AggregateResponse object
}
\description{
AggregateResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

