% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{Application}
\alias{Application}
\title{Application Object}
\usage{
Application(detailsUrl = NULL, name = NULL, packageName = NULL,
  version = NULL)
}
\arguments{
\item{detailsUrl}{An optional URI that can be used to link back to the application}

\item{name}{The name of this application}

\item{packageName}{Package name for this application}

\item{version}{Version of the application}
}
\value{
Application object
}
\description{
Application Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}

