% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{BucketByActivity}
\alias{BucketByActivity}
\title{BucketByActivity Object}
\usage{
BucketByActivity(activityDataSourceId = NULL, minDurationMillis = NULL)
}
\arguments{
\item{activityDataSourceId}{The default activity stream will be used if a specific activityDataSourceId is not specified}

\item{minDurationMillis}{Specifies that only activity segments of duration longer than minDurationMillis are considered and used as a container for aggregated data}
}
\value{
BucketByActivity object
}
\description{
BucketByActivity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

