% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{BucketByTime}
\alias{BucketByTime}
\title{BucketByTime Object}
\usage{
BucketByTime(durationMillis = NULL, period = NULL)
}
\arguments{
\item{durationMillis}{Specifies that result buckets aggregate data by exactly durationMillis time frames}

\item{period}{No description}
}
\value{
BucketByTime object
}
\description{
BucketByTime Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

