% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{BucketByTimePeriod}
\alias{BucketByTimePeriod}
\title{BucketByTimePeriod Object}
\usage{
BucketByTimePeriod(timeZoneId = NULL, type = NULL, value = NULL)
}
\arguments{
\item{timeZoneId}{org}

\item{type}{No description}

\item{value}{No description}
}
\value{
BucketByTimePeriod object
}
\description{
BucketByTimePeriod Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

