% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{DataPoint}
\alias{DataPoint}
\title{DataPoint Object}
\usage{
DataPoint(computationTimeMillis = NULL, dataTypeName = NULL,
  endTimeNanos = NULL, modifiedTimeMillis = NULL,
  originDataSourceId = NULL, rawTimestampNanos = NULL,
  startTimeNanos = NULL, value = NULL)
}
\arguments{
\item{computationTimeMillis}{Used for version checking during transformation; that is, a datapoint can only replace another datapoint that has an older computation time stamp}

\item{dataTypeName}{The data type defining the format of the values in this data point}

\item{endTimeNanos}{The end time of the interval represented by this data point, in nanoseconds since epoch}

\item{modifiedTimeMillis}{Indicates the last time this data point was modified}

\item{originDataSourceId}{If the data point is contained in a dataset for a derived data source, this field will be populated with the data source stream ID that created the data point originally}

\item{rawTimestampNanos}{The raw timestamp from the original SensorEvent}

\item{startTimeNanos}{The start time of the interval represented by this data point, in nanoseconds since epoch}

\item{value}{Values of each data type field for the data point}
}
\value{
DataPoint object
}
\description{
DataPoint Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a single data point, generated by a particular data source. A data point holds a value for each field, an end timestamp and an optional start time. The exact semantics of each of these attributes are specified in the documentation for the particular data type.A data point can represent an instantaneous measurement, reading or input observation, as well as averages or aggregates over a time interval. Check the data type documentation to determine which is the case for a particular data type.Data points always contain one value for each field of the data type.
}
