% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{Dataset}
\alias{Dataset}
\title{Dataset Object}
\usage{
Dataset(dataSourceId = NULL, maxEndTimeNs = NULL, minStartTimeNs = NULL,
  nextPageToken = NULL, point = NULL)
}
\arguments{
\item{dataSourceId}{The data stream ID of the data source that created the points in this dataset}

\item{maxEndTimeNs}{The largest end time of all data points in this possibly partial representation of the dataset}

\item{minStartTimeNs}{The smallest start time of all data points in this possibly partial representation of the dataset}

\item{nextPageToken}{This token will be set when a dataset is received in response to a GET request and the dataset is too large to be included in a single response}

\item{point}{A partial list of data points contained in the dataset, ordered by largest endTimeNanos first}
}
\value{
Dataset object
}
\description{
Dataset Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A dataset represents a projection container for data points. They do not carry any info of their own. Datasets represent a set of data points from a particular data source. A data point can be found in more than one dataset.
}

