% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{DataType}
\alias{DataType}
\title{DataType Object}
\usage{
DataType(field = NULL, name = NULL)
}
\arguments{
\item{field}{A field represents one dimension of a data type}

\item{name}{Each data type has a unique, namespaced, name}
}
\value{
DataType object
}
\description{
DataType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}

