% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{DataTypeField}
\alias{DataTypeField}
\title{DataTypeField Object}
\usage{
DataTypeField(format = NULL, name = NULL, optional = NULL)
}
\arguments{
\item{format}{The different supported formats for each field in a data type}

\item{name}{Defines the name and format of data}

\item{optional}{No description}
}
\value{
DataTypeField object
}
\description{
DataTypeField Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
In case of multi-dimensional data (such as an accelerometer with x, y, and z axes) each field represents one dimension. Each data type field has a unique name which identifies it. The field also defines the format of the data (int, float, etc.).This message is only instantiated in code and not used for wire comms or stored in any way.
}

