% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{Device}
\alias{Device}
\title{Device Object}
\usage{
Device(manufacturer = NULL, model = NULL, type = NULL, uid = NULL,
  version = NULL)
}
\arguments{
\item{manufacturer}{Manufacturer of the product/hardware}

\item{model}{End-user visible model name for the device}

\item{type}{A constant representing the type of the device}

\item{uid}{The serial number or other unique ID for the hardware}

\item{version}{Version string for the device hardware/software}
}
\value{
Device object
}
\description{
Device Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Representation of an integrated device (such as a phone or a wearable) that can hold sensors. Each sensor is exposed as a data source.The main purpose of the device information contained in this class is to identify the hardware of a particular data source. This can be useful in different ways, including:  - Distinguishing two similar sensors on different devices (the step counter on two nexus 5 phones, for instance)- Display the source of data to the user (by using the device make / model)- Treat data differently depending on sensor type (accelerometers on a watch may give different patterns than those on a phone)- Build different analysis models for each device/version.
}
