% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_functions.R
\docType{package}
\name{fitness_googleAuthR}
\alias{fitness_googleAuthR}
\alias{fitness_googleAuthR-package}
\title{Fitness
Stores and accesses user data in the fitness store from apps on any platform.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:12:46
filename: /Users/mark/dev/R/autoGoogleAPI/googlefitnessv1.auto/R/fitness_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/fitness.activity.read
\item https://www.googleapis.com/auth/fitness.activity.write
\item https://www.googleapis.com/auth/fitness.body.read
\item https://www.googleapis.com/auth/fitness.body.write
\item https://www.googleapis.com/auth/fitness.location.read
\item https://www.googleapis.com/auth/fitness.location.write
\item https://www.googleapis.com/auth/fitness.nutrition.read
\item https://www.googleapis.com/auth/fitness.nutrition.write
}
}

