% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{ListDataSourcesResponse}
\alias{ListDataSourcesResponse}
\title{ListDataSourcesResponse Object}
\usage{
ListDataSourcesResponse(dataSource = NULL)
}
\arguments{
\item{dataSource}{A previously created data source}
}
\value{
ListDataSourcesResponse object
}
\description{
ListDataSourcesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
