% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{ListSessionsResponse}
\alias{ListSessionsResponse}
\title{ListSessionsResponse Object}
\usage{
ListSessionsResponse(deletedSession = NULL, nextPageToken = NULL,
  session = NULL)
}
\arguments{
\item{deletedSession}{If includeDeleted is set to true in the request, this list will contain sessions deleted with original end times that are within the startTime and endTime frame}

\item{nextPageToken}{The continuation token, which is used to page through large result sets}

\item{session}{Sessions with an end time that is between startTime and endTime of the request}
}
\value{
ListSessionsResponse object
}
\description{
ListSessionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

