% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{MapValue}
\alias{MapValue}
\title{MapValue Object}
\usage{
MapValue(fpVal = NULL)
}
\arguments{
\item{fpVal}{Floating point value}
}
\value{
MapValue object
}
\description{
MapValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Holder object for the value of an entry in a map field of a data point.A map value supports a subset of the formats that the regular Value supports.
}
