% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{Session}
\alias{Session}
\title{Session Object}
\usage{
Session(activeTimeMillis = NULL, activityType = NULL, application = NULL,
  description = NULL, endTimeMillis = NULL, id = NULL,
  modifiedTimeMillis = NULL, name = NULL, startTimeMillis = NULL)
}
\arguments{
\item{activeTimeMillis}{Session active time}

\item{activityType}{The type of activity this session represents}

\item{application}{The application that created the session}

\item{description}{A description for this session}

\item{endTimeMillis}{An end time, in milliseconds since epoch, inclusive}

\item{id}{A client-generated identifier that is unique across all sessions owned by this particular user}

\item{modifiedTimeMillis}{A timestamp that indicates when the session was last modified}

\item{name}{A human readable name of the session}

\item{startTimeMillis}{A start time, in milliseconds since epoch, inclusive}
}
\value{
Session object
}
\description{
Session Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Sessions contain metadata, such as a user-friendly name and time interval information.
}
\seealso{
Other Session functions: \code{\link{users.sessions.update}}
}
