% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_functions.R
\name{users.dataSources.update}
\alias{users.dataSources.update}
\title{Updates the specified data source. The dataStreamId, dataType, type, dataStreamName, and device properties with the exception of version, cannot be modified.Data sources are identified by their dataStreamId.}
\usage{
users.dataSources.update(DataSource, userId, dataSourceId)
}
\arguments{
\item{DataSource}{The \link{DataSource} object to pass to this method}

\item{userId}{Update the data source for the person identified}

\item{dataSourceId}{The data stream ID of the data source to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/fitness.activity.write
\item https://www.googleapis.com/auth/fitness.body.write
\item https://www.googleapis.com/auth/fitness.location.write
\item https://www.googleapis.com/auth/fitness.nutrition.write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/fitness.activity.write, https://www.googleapis.com/auth/fitness.body.write, https://www.googleapis.com/auth/fitness.location.write, https://www.googleapis.com/auth/fitness.nutrition.write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/fit/rest/}{Google Documentation}

Other DataSource functions: \code{\link{DataSource}},
  \code{\link{users.dataSources.create}},
  \code{\link{users.dataSources.patch}}
}

