% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_functions.R
\name{users.sessions.update}
\alias{users.sessions.update}
\title{Updates or insert a given session.}
\usage{
users.sessions.update(Session, userId, sessionId, currentTimeMillis = NULL)
}
\arguments{
\item{Session}{The \link{Session} object to pass to this method}

\item{userId}{Create sessions for the person identified}

\item{sessionId}{The ID of the session to be created}

\item{currentTimeMillis}{The client's current time in milliseconds since epoch}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/fitness.activity.write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/fitness.activity.write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/fit/rest/}{Google Documentation}

Other Session functions: \code{\link{Session}}
}

