% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{Value}
\alias{Value}
\title{Value Object}
\usage{
Value(fpVal = NULL, intVal = NULL, mapVal = NULL, stringVal = NULL)
}
\arguments{
\item{fpVal}{Floating point value}

\item{intVal}{Integer value}

\item{mapVal}{Map value}

\item{stringVal}{String value}
}
\value{
Value object
}
\description{
Value Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Holder object for the value of a single field in a data point.A field value has a particular format and is only ever set to one of an integer or a floating point value.
}

