% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_objects.R
\name{ValueMapValEntry}
\alias{ValueMapValEntry}
\title{ValueMapValEntry Object}
\usage{
ValueMapValEntry(key = NULL, value = NULL)
}
\arguments{
\item{key}{No description}

\item{value}{No description}
}
\value{
ValueMapValEntry object
}
\description{
ValueMapValEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
